Add list orders command

+++ src/main/xyz/mcex/plugin/equity/database/Order.java
@@ -6,10 +6,10 @@ class Order
 {
   public final UUID playerUuid;
   public final int quantity;
-  public final int price;
+  public final double price;
   public final int rowId;
 
-  public Order(int rowId, UUID playerUuid, int quantity, int price)
+  public Order(int rowId, UUID playerUuid, int quantity, double price)
   {
     this.playerUuid = playerUuid;
     this.quantity = quantity;
